Create Generate Random password
we define a function generate_random_password() that takes an optional parameter length, which specifies the length of the password. The function creates a pool of characters using string.ascii_letters (lowercase and uppercase letters), string.digits (digits 0-9), and string.punctuation (special characters). It then randomly selects characters from this pool to form the password using the random.choice() function.